Add support for dynamic tests
Some modern test frameworks like KotlinTest, create their tests
dynamically and do not give anything besides the root TestDescriptor in
the TestPlan. For this reason, pitest completely fails to detect any
unit tests when used with recent versions of KotlinTest.

I've tried to fix it by adding a mode that detects dynamically
configured tests by executing the TestPlan and using a
TestExecutionListener to catch them.

Since there is no option to configure this mode with the current pitest
configuration, I'm checking if includedGroups contains the magic string
'@dynamic-tests' to determine whether to enable this mode.