support failing BeforeAlls
When a BeforeAll fails, jupiter generates a container, which never
registers as starting. Other engines may do something similar. To work
around this, and failing test identifier is now recorded even if it does
not identify as a test.

These failing units will still not appear to start, which may have
consequences that need to be examined.