update for pitest 1.9.0
Pitest 1.9.0 adds support for test plugins where the discovery
and execution of tests is not seperated (ie JUnit 5).

Previously each JUnit 5 test was executed twice during the coverage
stage, once to discover it, then again to run it for real.

This change takes advantage of the interfaces modified in 1.9.0 to
ensure that JUnit 5 tests are now discovered and run with one execution.